limitsAndRangeOf5

+++ src/main/java/ar/edu/unq/desapp/grupoa022022/backenddesappapi/service/serviceimpl/CryptocurrencyService.java
@@ -1,5 +1,6 @@
 package ar.edu.unq.desapp.grupoa022022.backenddesappapi.service.serviceimpl;
 
+import ar.edu.unq.desapp.grupoa022022.backenddesappapi.dto.CryptocurrencyRegister;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.Cryptocurrency;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.Intention;
 import ar.edu.unq.desapp.grupoa022022.backenddesappapi.model.Quote;
@@ -23,14 +24,22 @@ public class CryptocurrencyService implements ICryptocurrencyService {
     private IQuoteRepo quoteRepo;
 
     @Override
-    public Cryptocurrency create(String name) {
-        Cryptocurrency cryptocurrency = new Cryptocurrency(name);
-        return cryptocurrencyRepo.save(cryptocurrency);
+    public Cryptocurrency create(CryptocurrencyRegister cryptocurrencyRegister) {
+        Cryptocurrency cryptocurrency = new Cryptocurrency(cryptocurrencyRegister.getName());
+        Cryptocurrency savedCryptocurrency = cryptocurrencyRepo.save(cryptocurrency);
+        Quote quote = new Quote(savedCryptocurrency, cryptocurrencyRegister.getPrice());
+        quoteRepo.save(quote);
+        return savedCryptocurrency;
     }
 
     @Override
     public void delete(int id) {
-        cryptocurrencyRepo.deleteById(id);
+        if (cryptocurrencyRepo.findById(id).isPresent()) {
+            cryptocurrencyRepo.findById(id).get().getQuotes().forEach(quote -> {
+                quoteRepo.deleteById(quote.getId());
+            });
+            cryptocurrencyRepo.deleteById(id);
+        }
     }
 
     @Override
@@ -40,7 +49,9 @@ public class CryptocurrencyService implements ICryptocurrencyService {
 
     @Override
     public void deleteAll() {
-        cryptocurrencyRepo.deleteAll();
+        cryptocurrencyRepo.findAll().forEach(cryptocurrency -> {
+            this.delete(cryptocurrency.getId());
+        });
     }
 
     @Override
